+++ source/source.h
@@ -19,13 +19,13 @@ extern "C" {
 #endif
 
 EXPORT_API int file_init_source(void *ctx);
-EXPORT_API int file_read_data(void *ctx, char* buff, int64_t offset, int buf_size);
+EXPORT_API int64_t file_read_data(void *ctx, char* buff, int64_t offset, size_t buf_size);
 EXPORT_API void file_close(void *ctx);
 EXPORT_API void file_destory(void *ctx);
 
 EXPORT_API int bt_init_source(void *ctx);
 EXPORT_API int bt_media_info(void *ctx, char *name, int64_t *pos, int64_t *size);
-EXPORT_API int bt_read_data(void *ctx, char* buff, int64_t offset, int buf_size);
+EXPORT_API int64_t bt_read_data(void *ctx, char* buff, int64_t offset, size_t buf_size);
 EXPORT_API void bt_close(void *ctx);
 EXPORT_API void bt_destory(void *ctx);
