+++ source/torrent_source.h
@@ -66,7 +66,7 @@ public:
 	virtual bool open(void* ctx);
 
 	// 读取数据.
-	virtual bool read_data(char* data, uint64_t offset, uint64_t size, uint64_t& read_size);
+	virtual bool read_data(char* data, uint64_t offset, size_t size, size_t& read_size);
 
 	// 关闭.
 	virtual void close();
